+++ build_tools/circle/build_wheel.sh
@@ -8,11 +8,16 @@ function build_wheel {
     local ucs_setting=$3
 
     # https://www.python.org/dev/peps/pep-0513/#ucs-2-vs-ucs-4-builds
-    ucs_tag="m"
-    if [ "$pyver" = "3.8" ] || [ "$pyver" = "3.9" ]; then
-        ucs_tag=""
-    elif [ "$ucs_setting" = "ucs4" ]; then
+    ucs_tag=""
+    if [ "$ucs_setting" = "ucs4" ]; then
         ucs_tag="${ucs_tag}u"
+    fi
+
+    distutils_version=""
+    if [ "$pyver" = "3.12" ]; then
+      distutils_version="local"
+    else
+      distutils_version="stdlib"
     fi
 
     ML_PYTHON_VERSION=$(python -c \
@@ -23,7 +28,7 @@ function build_wheel {
 
     DOCKER_CONTAINER_NAME=wheel_builder_$(uuidgen)
 
-    ML_IMAGE="quay.io/pypa/manylinux_2_24_${arch}:2021-08-29-ad54d32" # `latest` as of 2021-08-30
+    ML_IMAGE="quay.io/pypa/manylinux_2_28_${arch}:2023-10-07-c1e05d1" # `latest` as of 2023-10-09
     PMDARIMA_VERSION=`cat ~/pmdarima/pmdarima/VERSION`
 
     docker pull "${ML_IMAGE}"
@@ -33,6 +38,7 @@ function build_wheel {
         -v `pwd`:/io \
         -e "PYTHON_VERSION=${ML_PYTHON_VERSION}" \
         -e "PMDARIMA_VERSION=${PMDARIMA_VERSION}" \
+        -e "SETUPTOOLS_USE_DISTUTILS=${distutils_version}" \
         "${ML_IMAGE}" "/io/build_tools/circle/dind/build_manylinux_wheel.sh"
     sudo docker cp "${DOCKER_CONTAINER_NAME}:/io/dist/." "${_root}/dist/"
     docker rm $(docker ps -a -f status=exited -q)